requirements.txt
mcp
httpx
python-dotenv
pydantic
uvicorn